Put a rule in to generate a valid ActiveThorns file.
This will call BuildActiveThorns.pl from the lib/sbin directory to
generate a list.

It then displays the list, and asks the user if they want to edit it.
If they say yes/y/YES/Y it calls an editor.  Vi by default, but will use
value of the $EDITOR environment variable if that is set.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@75 17b73243-c579-4c4c-a9d2-2d5706c11dac