New library - libCactusBindings.a created by make.configuration.
The sources for this are automatically generated by config_parser.pl
and should contain everything needed to bind the flesh and the
thorns.

Specifically have six functions called from the flesh

CCTK_BindingsParametersInitialise
CCTK_BindingsParameterSet
CCTK_BindingsParameterGet

CCTK_BindingsVariablesInitialise

CCTK_BindingsScheduleInitialise
CCTK_BindingsScheduleRegister

All the initialise ones are of type int ()(void)

CCTK_ParameterSet is of type int ()(const char *, const char *)
and takes a parameter specification and a parameter value, as taken
from the parameter file, and sets the parameter accordingly.

CCTK_BindingsParameterGet is of type int()(const char *, void **)
and returns the value of a parameter - the rturn value corresponds to
the type of the parameter.

CCTK_BindingsScheduleRegister is of type int()(const char *)
and will perform registration at, e.g. Startup, or RFR, or
any other point we add in future.  Thinking about it, it needs
to also take a void * if it's going to do rfr init stuff - watch this space...
8-)

The c source files for these are put in the $(TOP)/bindings directory by
config.ccl, each subsystem in its own directory.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@82 17b73243-c579-4c4c-a9d2-2d5706c11dac