Rationalised the overloading of functions.  Goodbye RegisterKeyedFunction, hello Overloadable !
Now each subdirectory has one header file (in the include dir) which
defines the overloadable functions.  Macros are then applied to the
header file to produce functions of the form

CCTK_OverloadFoo(... (*func)(...))

to overload the function, and to create prototypes, dummy functions, and
if statements which can be used to assign dummy (or other default) functions
to any overloadable function which hasn't been assigned.

The resulting functions have names of the form CCTK_Foo()

As a byproduct of this the communication layer should work, 'though it's
all dummy functions there at the moment, but at least you can call the functions.

I've put OutputGH and OutputGroup in the IO layer, but can't remember
precisely what we had decided as the canonical set of overloadable functions
for IO.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@206 17b73243-c579-4c4c-a9d2-2d5706c11dac