Added two new options...
To rebuild a configuration run gmake followed by the name of a configuration suffixed by -rebuild e.g. Linux-rebuild. This forces the CST to be rerun.

To reconfigure a configuration run gmake followed by the name of a configuration suffixed by -reconfig e.g. Linux-reconfig. This reruns the configuration scripts.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@290 17b73243-c579-4c4c-a9d2-2d5706c11dac