Can now pass options to the configure script
use

make <config> options=<option-file-name>

when making a new configuration.

<option-file-name> has the format of

    keyword value
or  keyword = value

Note that even though it will allow a vlue to be blank, the configure
script itself will ignore this, so it's not so easy to tell it to ignore,
say, your f90 compiler.

The configure script now uses the f90 compiler for f77, or the f77
compiler if there is no f90 compiler.

The configure.pl script will produce a dummy fortran_name perl script if
there is no fortran compiler.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@319 17b73243-c579-4c4c-a9d2-2d5706c11dac