Now the presence of old parameter header files won't confuse the system.
Note that the dependencies may not pick up that things now depend on a
different set of files, so it may still be necessery to do a gmake ..-clean
to force files to be rebuilt - e.g. changing all parameters from private
to protected won't be picked up by the current dependency checking
mechanism.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@392 17b73243-c579-4c4c-a9d2-2d5706c11dac