Added CCTK_ParamWarn to give a warning about parameters and add one to the count of parameter warnings. Use this in any routine registered at ParamCheck to give a warning which should stop the code.
This has a macro CCTK_PARAMWARN("blah")
which adds on the thorn name

Changed CCTK_Info to also take the thorn name, and there is now a
macro CCTK_INFO("blah") to automatically add it.

Also added CCTKi_Paramfinalise which is called by flesh after param
check to stop the code if there are parameter problems.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@592 17b73243-c579-4c4c-a9d2-2d5706c11dac