New names for parameter groups:
public -> global

protected -> restricted

private -> private

friend -> shares (although we would like a better word for this)

Can now use USES as well as EXTENDS (although both of them are
optional ... perhaps they shouldn't be?).

I would like to be able to get rid of the { } after USES but I
couldn't manage it so far.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@632 17b73243-c579-4c4c-a9d2-2d5706c11dac