Stuff for the ActiveThorns parameter.
Now need to specify ActiveThorns before any other parameters to list those
thorns which are active.  No parameters or startup or rfr routines are
available from inactive thorns.

Note that this isn't quite working yet but I hope to track it down by morning.

In the meantime either don't update or be prepared for things to be slightly
broken.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@645 17b73243-c579-4c4c-a9d2-2d5706c11dac