Implements the generic CCTK_Interp() routine and registration of interpolation operators.
Call the Interp with arguments

int CCTK_Interp (cGH *GH,
                 int operation_handle,
                 int nPoints,
                 int nCoords,
                 int nInFields,
                 int nOutFields,
                 ...)

whereas in the variable argument list there are passed (int this order):

  - coordinate arrays of points to interpolate at (void *[])
  - array of variable types for each coordinate array (int [])
  - array of input fields, given as grid functions indices (int [])
  - array of output fields (void *[])
  - array of variable types for output field (int [])

The Fortran CCTK_Interp gets in addition to those as the second parameter
a pointer to an integer where the returncode is stored.


See the packages/CactusPUGH/interp/src/interp_demo.F for an example
of using the interp.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@692 17b73243-c579-4c4c-a9d2-2d5706c11dac