Changed the interface for CCTK_Interp which is now more general because pointers to input fields are passed instead of variable indices. This way you can also interpolate on local fields.
int CCTK_Interp (cGH *GH,
                 int operation_handle,
                 int nPoints,
                 int nDims,
                 int nInFields,
                 int nOutFields,
                 ...)

whereas in the variable argument list there are passed (int this order):

  - dimensions of the underlying grid (int [nDims])
  - coordinate arrays of points to interpolate at (void *[nDims])
  - array of variable types for each coordinate array (int [nDims])
  - origin of the underlying grid (<type of coordinate fields> [nDims])
  - deltas of the underlying grid (<type of coordinate fields> [nDims])
  - array of input fields (void *[nInFields])
  - array of variable types for input fields (int [nInFields])
  - array of output fields (void *[nOutFields])
  - array of variable types for output fields (int [])

The Fortran CCTK_Interp gets in addition to those as the second parameter
a pointer to an integer where the returncode is stored.



For interpolation of grid functions you can use CCTK_InterpGF
which has the same interface as the old CCTK_Interp routine.
Here you just pass the GF indices instead of the pointers.
Since GFs are a special case of arrays, this routine uses
the internal description of the underlying grid and simply
calls CCTK_Interp.


Interpolation operators have to implement the interface for the
most general CCTK_Interp.


See the packages/CactusPUGH/interp/src/interp_demo.F for an example
of how to use both CCTK_Interp and CCTK_InterpGF.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@708 17b73243-c579-4c4c-a9d2-2d5706c11dac