Added lower and upper bound coordinate registration for each coordinate on each GH. The relevant functions are
CCTK_RegisterCoordRange(GH,lower,upper,"name")

(this is done in CartGrid3D now after the coords are calculated)

then to get at the values

CCTK_CoordRange(GH,lower,upper,"name")

Note that in C lower and upper are pointers to the values.

This only all works for CCTK_REAL coords at the moment.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@793 17b73243-c579-4c4c-a9d2-2d5706c11dac