Added yet another make variable.
Now the make stuff required for the final linking of libraries is
bundled up into

GENERAL_LIBRARIES

which defaults to

$(LIBDIRS:%=-L%) $(LIBS:%=-l%)

which should work on all unices.

For NT I have made it

$(LIBS)

and LIBDIRS has no use.  NT stuff needs to specify the full path and name of
each library in the LIBS variable.  Maybe one day we'll find a neater way !

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@874 17b73243-c579-4c4c-a9d2-2d5706c11dac