Configuration script for parallel HDF5 (configure option HDF5=PARALLEL)
It checks that it is configured together with MPI and that HDF5_DIR
really points to a parallel version of HDF5.

The tricky thing with MPI is that the HDF5 lib should have been compiled
with the same MPI version that is used for Cactus. Don't know how
to check this for now.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@916 17b73243-c579-4c4c-a9d2-2d5706c11dac