Added a make.code.defn.
Added an extra data field to the traversal calls so that things like the GH can be passed
through.

Schedule helper routines now return an error code.  Currently only the ones from
the entry and the while_list routines are used, and both indicate if processing of
that group/function whoculd be done.

The while_list is checked first, and if that returns success, the entry one is called,
if that returns success, the loop is entered, and that loop exits when while_loop returns
false.  Finally the exit function is called unless the entry function was never called.

If these functions aren't provided (i.e. they are NULL), the code does the main loop once.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@937 17b73243-c579-4c4c-a9d2-2d5706c11dac