More updates to timer stuff.
Now to get info from a timer, one must create a timer info function:

info = CCTK_TimerCreateInfo();

Get the data

CCTK_TimerGet(timer, info);

Query the fields in the info structure, and then either query another timer
with the same info structure, or destroy the info structure.

CCTK_TimerDestroyInfo(info);

The info structure contains the number of values, and an array of
these values, along with header info, units, data types and values.

I think this is a fairly complete interface, but if a better one can be
found, this may change.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@1083 17b73243-c579-4c4c-a9d2-2d5706c11dac