Tom's changes to add configuration option
gmake <config>[-config] THORNLIST=<file name>

to use a given file as the thorn list for that configuration.

This needs to be tweaked to do something if the file to be copied doesn't
exist, but I'm still wondering whether to exit with an error, prompt for
another filename or do the default of making the thornlist from all the
available thorns. Also, it might be useful to give a warning if it
includes thorns which are not available, and also to still give you
a chance to edit the file.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@1261 17b73243-c579-4c4c-a9d2-2d5706c11dac