Added
SYNC[HRONISE]: <group list>

as an option to the schedule.ccl.

This will be used to tell the flesh or the driver to synchronise
these groups after the function has been executed.

Tom

NOTE:  You will need to do a

gmake <config>-rebuild

after this update as the bindings stuff has changed.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@1318 17b73243-c579-4c4c-a9d2-2d5706c11dac