Now, subject to review, you can put
OPTIONS:...

in a schedule block to pass miscellaneous options to the schedular.

Currently the only option is "Global" which sets a flag on the
FunctionData structure passed to the call function.  This option
only makes sense for a scheduled function, not a schedule group,
though this may change in the future.  It is used to indicate to
the calling function that this function should only be called once
as it performs global operations.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@1325 17b73243-c579-4c4c-a9d2-2d5706c11dac