Added
C_POSTPROCESSING
CXX_POSTPROCESSING
F77_POSTPROCESSING
F_POSTPROCESSING
F90_POSTPROCESSING

as things a known-architecture can set.  This allows object files to be
moved and things like that.

Also added the ability of the architecture files to write to a file calles

make.arch.defn

which is included at the bottom of make.code.defn.  Thus, for example,
an architecture requiring mpicc and its ilk for compiling with MPI
can replace the compilers at this point, but still allow the rest
of the configuration to use the normal compiler.  Use with caution.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@1438 17b73243-c579-4c4c-a9d2-2d5706c11dac