Tidied a bit.
Added creation of a new file

configs/<config>/config-info

which stores info about the configuration.

Currently it stores the name of the configuration, the flags passed to make,
the date of the configuration, the host the configuration was created on,
and the error return code of the configure script.

Note that the flags are in the form

<actual make flags> -- <other flags>

and normally will just be

s

which is from the implicit -s when SILENT=yes (the default).

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@1570 17b73243-c579-4c4c-a9d2-2d5706c11dac