Support for mpich-1.2.0 .
Note that the new directory layout in this version of mpich means that
it no longer needs MPICH_ARCH or MPICH_DEVICE, although if you specify
them it increases the range of directories to look for mpich in.

You still need to use MPICH_DEVICE to get the special libraries for
globus or myrinet.

It still works with the older MPICH version.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@1594 17b73243-c579-4c4c-a9d2-2d5706c11dac