Changed checkout description to use GetCactus and ThornLists, and did a bit of spellchecking.
If anyone knows how to change the definition of the Lentry macros so that
if there is no text for the item it doesn't give a new line, I would like to know it, since I couldn't work out how to change it. (For example, the list of supported architectures looks bad using Lentry).


git-svn-id: https://svn.cactuscode.org/flesh/trunk@1909 17b73243-c579-4c4c-a9d2-2d5706c11dac