Initial implementation of Function Aliasing for thorns.
This is described on a Spec on the web pages at the moment, and will move to the documentation
once it is tested and extended a bit more. If you want to have a look at it in action, checkout
the thorns TestFunctions1A, TestFunctions1B and TestFunctions2 which will be in AlphaThorns in
a few minutes time.

At the moment it is only possible to use Function Aliasing for C functions
which use and return Cactus data types.

You will need to rebuild your configurations once you update, but apart from that these changes
shouldn't affect anyone ... if you see any problems please let us know.

Gabrielle


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2060 17b73243-c579-4c4c-a9d2-2d5706c11dac