Added Util_asprintf,Util_asnprintf,Util_snprintf,Util_vsnprintf.
asprintf is like sprintf, but the first arg is a char ** and should basically be
the address of a null pointer.  The routine allocates enough space and passes back
the new buffer pointer.

asnprintf is as above, but the buffer need not be null.  The thirs argument is the
size of the current buffer, and if the string would be larger than that, the current
buffer is freed and a new buffer big enough is allocated.

snprintf and vsnprintf are pretty standard, except we don't have them on all platforms
so I hunted down a file on the internet, written by Patrick Powel (papowell@astart.com),
to give us a portable version of the functionality.  It may not be entirely up to
scratch with floating point - I've not tested it too thouroughly yet.

This should enable us to make the web-server stuff a bit more secure against buffer-overruns.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2209 17b73243-c579-4c4c-a9d2-2d5706c11dac