Added function CCTK_WriteFiles which compares file against old version and only overwrites if the new version is genuinely new.  Now CCTK_CreateFile saves a list of files it has created.
Changed the stuff in configure.in to use CCTK_CreateFile to create the
architecture and extras files and to always use CCTK_WriteLine to write to
a file.

Now if you reconfigure with precisely the same config options you don't
end up rebuilding the whole configuration if you have MPI or HDF5.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2248 17b73243-c579-4c4c-a9d2-2d5706c11dac