Preliminary support for 'vector' Grid Variables -
defined as

<type> <group_name>[<size>] ...

as per a normal 1 variable group except with an extra argument.  This
creates a vector of GVS, with an extra index (the last one for a GA or GF)
corresponding to the element in the vector.  For scalars this creates an array
similar to a DISTRIB=CONSTANT GA. <size> may be either an integer or an
integer-valued parameter (which must not be steerable).

So, fortran sees

REAL foo[param] TYPE=GF DIM=3

as

CCTK_REAL foo(nx,ny,nz,param)

and in C it should be accessed as foo[CCTK_GFINDEX3D(cctkGH,i,j,k)*l]

When we have COMPACT groups you will also be able to make the extra index
the first one rather than the last one.

Please treat this as volatile until it is documented, and let me know if you are using
it.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2393 17b73243-c579-4c4c-a9d2-2d5706c11dac