Added fortran wrapper for CCTK_ParameterValString(). Use it as
  call CCTK_ParameterValString (ierror, parameter, thorn, value)

with parameter, thorn, value being fortran strings, the last one holding
the actual stringified value of the parameter.
ierror must be of type CCTK_INT and holds the return code of this wrapper
where 0 stands for success, -1 for 'parameter doesn't exist', -2 for
'fortran string buffer is too short to hold the value.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2446 17b73243-c579-4c4c-a9d2-2d5706c11dac