Changed CCTK_TimerPrintData and CCTK_TimerPrintDataI to take two arguements, so you can choose both a timer and a clock to print. Passing a NULL argument for a string, or a -1 for the index versions will cause all the timers/clocks to be used.
Also added new functions

CCTK_NumClocks
CCTK_ClockName
CCTK_ClockHandle

which I needed for this. Not sure that I really needed the ClockHandle
one, but it is useful anyway to test if a clock is registered.

Maybe these functions should also take something like a file descriptor,
so that the information can be easily printed to a file, or to the HTTPD
pages.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2547 17b73243-c579-4c4c-a9d2-2d5706c11dac