ThornWriters.tex    section B8.3 "Interpolation Operators"    * clarify description of how interpolation operators are registered    * add descriptions of new proposed interpolation APIs for grid arrays      and for local "nonuniform" and "warped" grids, all commented-out      (see below)
FunctionReference.tex
* add descriptions of new proposed interpolation APIs for grid arrays
  and for local "nonuniform" and "warped" grids, all commented-out
  (see below)

With this checkin, there are now 4 new interpolation APIs documented here,
but 3 of them are commented out in the latex source with the string
%notyet^I   (that's a %notyet at the start of the line followed by a
tab character).  This way the typeset manual doesn't describe
not-yet-implemented APIs.  At the moment this is the only place where
the string "notyet" appears in *.tex in this directory, so grepping
for this string will find all the commented-out text.

The new interpolation APIs and their commented-out-or-not status are:
	%notyet	CCTK_InterpGridArrays
	CCTK_InterpLocalUniform
	%notyet	CCTK_InterpLocalNonUniform
	%notyet	CCTK_InterpLocalWarped

I will be away for the next 2 weeks; talk to Thomas Radke or Tom Goodale
or Ian Hawke for more info about the new interpolators.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2635 17b73243-c579-4c4c-a9d2-2d5706c11dac