Two new overloadable routines
CCTK_GroupStorageIncrease(cGH *GH, int n_groups,const int *groups,const int *timelevels, int *status)
CCTK_GroupStorageDecrease(cGH *GH, int n_groups,const int *groups,const int *timelevels, int *status)

These should increase or decrease the number of timelevels with storage enabled to the
number specified in the timelevels array for each group in the groups array.  The status
array is optional, and if not NULL should contain the previous number of timelevels enabled
for that group.  A timelevel value of -1 implies enable/disable all timelevels for that group.

The Increase function should never decrease the amount of storage, and the Decrease one never
increase it, so it is possible to use Enable to ensure a certain minimum number of timelevels
are available without problem, and to pass the status return from Enable in as the timelevel
array of Decrease to restore the previous state of the system.

The defaults for these functions currently call the old Enable and Disable GroupStorage
routines if they have been overloaded, similarly the defaults for the old routines have been
changed to call the new routines if they are overloaded, passing -1 as the timelevel value.

The old routines will disappear eventually once all drivers are converted to the new routines.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2693 17b73243-c579-4c4c-a9d2-2d5706c11dac