Changes to parse and validate a timelevel specifier for storage. If you have a group with more than one timelevel it is now an error not to specify in your schedule.ccl file how many timelevels to enable.
Note that the information is still not passed into the flesh, this
will happen soon...

This change should not affect anyone who is not using timelevels.
Everyone who is will need to update their schedule.ccl files, the syntax is

storage: group[<number>]

to enable storage for <number> timelevels, where <number> must be between
1 and the number of timelevels specified in the interface.ccl file,
inclusive.

See CactusWave for examples.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2724 17b73243-c579-4c4c-a9d2-2d5706c11dac