Support for the C99 restrict keyword.  This is dealt with in the same way as the const keyword - i.e. if it exists we do
#define restrict restrict

and if it doesn't exist we do

#define restrict

so it is safe to start using the restrict keyword in your C files, they will
still compile with compilers which don't support restrict.

Note that if you had a local variable called 'restrict' in C you will get an
error message, however you would have received one anyway when compiling with
a C99 compliant compiler.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2796 17b73243-c579-4c4c-a9d2-2d5706c11dac