TAGS tables for groups.
You can now do

TAGS='foo=... bar=...'

in a group definition in your interface.ccl files.

The intention is that this can be used to assign tensor types and coordinate
systems to gorups.

This is implemented as a table, the handle of which you can get by calling

CCTK_GroupTagsTable

or

CCTK_GroupTagsTableI

and query using the normal table operators.

Note that currently the Util_TableCreateFromString function, which is used to generate
the table, doesn't support setting of string members.  This will be changed soon.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2845 17b73243-c579-4c4c-a9d2-2d5706c11dac