Added support for a config file for testsuites, where IO thorns can provide information about the data files which they produce, and thorns providing testsuites can provide information about the tests they are supplying. The syntax for the config file is not finalised yet, and will be subject to change depending on the feedback we get.
The config file is called "config" and must be located in the "test"
directory of a thorn. Right now, the only line recognized is

EXTENSIONS dat xl yl zl

which should be added by any thorn providing output files eg with extensions
dat xl yl zl which should be tested.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@2864 17b73243-c579-4c4c-a9d2-2d5706c11dac