Allow routines scheduled within groups triggered in analysis to be run if the enclosing group is triggered.  The logic is now that if a group is triggered at analysis, all member routines or groups of that group will be automatically triggered unless they in turn have a trigger, in which case that overrides the trigger.
E.g.

schedule TestSchedule_CalcOne AT Analysis before TestSchedule_Analysis
{
  storage: testschedule::one
  trigger: testschedule::one
  lang: C
} "Calculate the value of one"

schedule group TestSchedule_Analysis at ANALYSIS
{
  trigger: testschedule::two testschedule::three
} "TestSchedule top level analysis group"

schedule TestSchedule_CalcTwo IN TestSchedule_Analysis
{
  trigger: testschedule::two
  storage: testschedule::two
  lang: C
} "Calculate the value of two"

schedule GROUP TestSchedule_Analysis2 in TestSchedule_Analysis after TestSchedule_CalcTwo
{
  trigger: testschedule::three
} "TestSchedule second level analysis group"

schedule TestSchedule_CalcThree IN TestSchedule_Analysis2
{
  trigger: testschedule::three
  storage: testschedule::three
  lang: C
} "Calculate the value of three"

will call the appropriate routines to calculate one, two or three if these are selected for
output.

As before, output is done on any trigger variable or variable group when a routine with
the variable as a trigger exits.  So in the above example "three" is output when CalcThree
exits.  If this is not desired, the trigger and storage for "three" could have been placed
on the enclosing schedule group, and then the routine would have been called even if output
was only requested for "two" and not for "three".

This fixes PR 694, and helps with PR 844.

Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@3188 17b73243-c579-4c4c-a9d2-2d5706c11dac