Fix a memory leak in Util_TableCreateFromString():
This routine first tries to create a table (if this fails it returns
an error code).  Then it tries to add entries to the table via
Util_TableSetFromString().  If this fails it previously returned
without destroy the table, thus leaking memory.  The fix is to
destroy the table before returning the error code.

This fixes bug Cactus/1561.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@3317 17b73243-c579-4c4c-a9d2-2d5706c11dac