Changed to using the cygwin version of Perl under cygwin.  This has the translation of CCTK_HOME to a windows format from the master Makefile, and forced the addition of a new Make macro
TRANSLATE_DIRS

which is used in various places in Make to translate things like

/cygwin/f/...

into

f:\...

This is a 'call'able macro in Make, and defaults to $(1).

There is a new subroutine in RunTestUtils which serves the same purpose.

This change should not affect anyone not using Cactus on Windows.  If you
are using Windows you will need to reconfigure, making sure that the Perl
you use is the cygwin one and not the ActiveState one.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@3742 17b73243-c579-4c4c-a9d2-2d5706c11dac