Escape percent signs for latex.
Do not escape backslash-quote any more, because this sequence cannot
be created.

Add comments stating that it does not make sense to change the case of
strings after they have been cleaned for latex.

Do not fix the places where this happens, because things seem to work
and a release is imminent.  Submitted PR 1832.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@3850 17b73243-c579-4c4c-a9d2-2d5706c11dac