Added tracking of options for thorn configuration scripts.  Now a thorn may specify
OPTIONS a b c

in a PROVIDES block in its configuration.ccl and these options will be
tracked and stored in the configuration's config-info file.

The config-info file, and any options file specified on the command line
are now used to setup the environment before a configuration script is
run by the CST.

Note that the only options stored to the config-info file by the CST are ones
specified in the configuration.ccl files, and configuration scripts are
invoked each time the CST is run, so any option used by a thorn's configuration
script MUST be mentioned in the configuration.ccl file to be remembered the
next time the CST is invoked.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@3855 17b73243-c579-4c4c-a9d2-2d5706c11dac