This is a partial fix for bug Cactus/1908   scheduling error neither aborts Cactus run nor gives useful error message so scheduling errors now abort the Cactus run.
In particular, this commit changes the error-handling strategy of
  CCTKi_DoScheduleSortAllGroups()
  ScheduleSortGroup()
from "print a message to stderr, then return error code"
to CCTK_VWarn(CCTK_WARN_ABORT, ...), i.e.
"print a message to stderr, then abort the Cactus run".

The CCTKi_DoScheduleSortAllGroups() message is the same one as before.
The ScheduleSortGroup() message is changed to include the group name.


(N.b.  the other part of Cactus/1908, that the schedule-error
message is uninformative, is *not* fixed by this commit.)


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4016 17b73243-c579-4c4c-a9d2-2d5706c11dac