Add '-Mx,125,0x200' to F77FLAGS and F90FLAGS when using Portland compilers. This is a workaround for calling C varargs functions from Fortran.
Thanks to Roberto De Pietri who had reported this problem to the PGI developers.
Here is the full explanation he got from them:

======================================================================== ==
Joined: 30 Jun 2004
Posts: 99
Location: Portland Group

PostPosted: Wed Sep 29, 2004 9:56 am    Post subject: Problem only on  Opterons     Reply with quote

This is a problem that only occurs on Opterons when using Fortran to C
with varargs.  When you compile a C program on Opteron using varargs,
the EAX register is expected  to contain the number of XMM registers
that will be used for the varargs. When you  call a C function with
varargs from Fortran the EAX register is not set since its not needed
in Fortran. Since EAX contains garbage, the C varargs becomes
corrupt.

To work around this problem is to use the x flag "-Mx,125,0x200" which
will set EAX to zero before every call. Note that we generally do not
release x flags to the public since they are for the compiler's
internal use, new, or experimental. So the caveat to using this
flag is that it has not been properly QA'd and is subject to change.
However, I believe in this case we plan to give it a name in a
future  release.
========================================================================


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4020 17b73243-c579-4c4c-a9d2-2d5706c11dac