Effects a fix for PR 1343 and a partial fix for PR 989
This was discussed on CactusMaint Wed, 27 Apr 2005 as
      Patch for PR 1913: ThornGuide and Latex errors
(Note the 1913 seems to be a typo--it should have been 1343)

Files changed: Makefile, doc/ThornGuide/Makefile

Affects targets: ThornGuide, ThornGuide.pdf, <config>-ThornGuide
      UsersGuide, ReferenceManual, MaintGuide

Main purpose:  To give the user a sensible error message in case something
      goes wrong with latex'ing, pointing them to the appropriate latex
      log files, and stopping the build.

Also, made on-screen messages a little clearer as to what is going on.

Since the patch was posted, these additions were suggested and implemented.

      * Instead of copying the ps files to the doc directory,
        they are now moved (more efficient)

      * If ThornList is missing, make <config>-ThornGuide now complains.

      * Now also greps for LaTeX Warning, and complains if it finds one.

Tested: Ran all targets on functioning CCL files, and on one with bad
      embedded LaTeX (a missing right-curley-bracket will do it).

To Do:
      1) make ThornGuide.pdf still doesn't work because
         pdflatex doesn't take eps files.  Would ps2pdf be a suitable
	 alternative?  (PR 1937)

      2) Erik and Jonathan would both like to see all latex warnings
	 about preferences worked out.

      3) Scripts that parse the CCL files still fail silently, resulting
         in incomplete documentation.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4061 17b73243-c579-4c4c-a9d2-2d5706c11dac