The perl cpp does not handle empty macro arguments correctly.  When the last macro argument is empty, it thinks there is one argument too few.  That is, the following code
#define M2(a,b)
M2(x,)

leads to an error message stating that M2 was used with only a single
argument.  This is wrong, as it is called with two arguments, where
the second argument is the empty string.  After reading the GNU cpp
manual, I think that the ANSI standard requires that empty arguments
must be handled.

The problem is that the routine SplitArgs that decodes the macro
arguments cannot distinguish between an empty last argument and no
argument, as both are represented by an empty string.  I change the
routine so that an empty string always means "empty argument".  This
means that SplitArgs cannot be called any more if there are no
arguments -- but this does not matter.  I place a corresponding if
statement around the call.

Empty macro arguments may cause problems on some system.  A later
patch will produce warnings or errors for them.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4179 17b73243-c579-4c4c-a9d2-2d5706c11dac